Committing a standalone version to be able to run QA checks separate from the hook, but using the same concept. (and same checks obviously)
The script qachecks2 is a *lua* script which requires the standalone version
of lua 5. If you're into QA checks, you should have that anyway, since they
are written in lua to be able to interface cleanly with monotone. The hook
based ones use the embedded lua interpreter in monotone, so the checks during
commit do not need the standalone lua) Further requirement is the lfs lua
library for some filesystem functions. Again, if into qa checks, this
shouldn't be a problem

This script reads files from stdin and passes those files into the QA checks.
This is a bit different than the qachecks(1) script which would just take the
files from the current directory (recursively) and run the checks on them.

The typical usage would be inside a working dir and do:

	mtn ls known <path> | qachecks2

but the usage is not limited to that obviously. Any input of file-lists on
stdin is good enough. There is one format restriction, which is that the files
on the input should be specified with their location relative to the monotone
repository.

Suppose you have a file critical_files.list which contains:
html/includes/xarCore.php
html/includes/xarTemplate.php

then:
	cat critical_files.list | qachecks2 
	
will run the checks on just those two files.

Further TODO:
- only tested on OSX (linux should be no problem, windows might be tricky, no
  idea)
- it's very, very minimal on the checking of errors
- it's optimized for speed (with the exception of doing line by line checking
  for braces) not for readability (TODO: if someone can make those true brace
  checking work, based on full content matching, that would be great, since
  that is much faster)
- the PHP fudging to check syntax is evil. it's php calling a php script
  (obviously) which calls php on the file to be syntax-checked, just to
  suppress output, because monotone can not do that in a hook using
  os.execute('something') it's dirty, but it works (php5, osx)
- TODO: xml parsing isnt done yet, lua has no xml parser, could use php xml
  (one of the gazillion ones) or os provided xmllint (which is far better, but
  requires an extra program)

Probably lots more, but we can at least clean out some duplication in the qa
check area now

	* bin/qachecks2: new file
	* bin/qachecks2: set attribute (mtn:execute)