Multiple authkeys:
The changes are localized to includes/xarSecurity.php and includes/xarSession.php.

Xaraya doesn't actually store the authid, it stores a random value which is used to compare against a returned authid (one way cipher, like password handling).  The changes I made now store an array of random values, each associated with a timestamp.  The format is [timestamp]-[random value], ie 1204762826-1784614311

The security file has changes to xarSecGenAuthKey() and xarSecConfirmAuthKey().

Every call to xarSecGenAuthKey adds another index to the array of stored values.

A key is removed when it's age is greater than the session activity timeout.

In order to prevent session data overflow, only the most recent 64 keys are maintained.  This should allow for a decent amount of activity.  (I expect some crazy module somewhere calls for a new key for each item in an item list... keeping 64 keys should avoid a complete recycling of the values).

The session file was modified only to change the value stored when the session is initialized, mainly as a failsafe, since every call to generate a key add another one... the initial value is unlikely to ever be used.