Bug 722
Implement new option in adminmenu block to show all module menus, not just
the current one.

This change only sends the module menu to the block template, which is
unchanged.  The block itself makes no effort to implement special behavior or 
handling of the non-current module menus; for now, that is up to the theme 
(to which Xaraya_Classic isn't particularly favorable).