Fix for bug 1990 and its duplicates.
Mask and privilege instances are now stored and handled internally as a
serialized array.  Instance values may now be passed as 0-indexed arrays or as
the familiar colon-delimited string.  Array is now the preferred format, and
is used throughout the core, except in <xar:sec> tags.

The historical colon-delimited format is now only used when the instance is
presented to the user.  The mask and privilege classes handle the conversion
between serialize() and colon internally.  Added xarMask::getInstanceDisplay()
to easily access colon format.

DATABASE CHANGE:
    [prefix]_privileges.xar_instance now varchar 255
    [prefix]_security_masks.xar_instance now varchar 255

Found and fixed a lingering hardcoded reference to .xd templates in
xarException.php.  Changed several references to .xd in template text.

Replace <font> with <span> in privileges/xarprivileges.php.